Renamed public-stats to public
Except in cases where it would cause an identifier named `public` as this is forbidden in strict mode.
